Next step towards: - All major terminal settings are now also available when creating a new terminal as well as when   calling YAT from the command line (feature requests #3162725 and #3391925, related to feature   request #2991946 listed below)
> Almost there...

+++ MKY/MKY/IO/StreamEx.cs
@@ -34,7 +34,7 @@ namespace MKY.IO
 		/// <summary>
 		/// Result when reading the stream's end.
 		/// </summary>
-		public const long EndOfStream = -1;
+		public const int EndOfStream = -1;
 
 		/// <summary>
 		/// Result when stream length cannot be retrieved from a stream, e.g. if the stream is not seekable.
